chore: 열거형 변환 메소드 추가 #6
